backend/requirements.txt
transformers
google-genai
torch
python-dotenv
fastapi
uvicorn
pydantic
python-multipart
pdfplumber
python-docx
